completed hot collections section, added skeleton loading state with shimmer effect
