+++ src/index.css
@@ -11,6 +11,12 @@
 @import "./css/colors/scheme-01.css";
 @import "./css/styles/coloring.css";
 
+/*
+
+HOT COLLECTION 
+
+*/
+
 .slider__wrapper {
   position: relative;
 }
@@ -44,9 +50,55 @@
 }
 
 .slider-next-btn {
-  right: 8px;
+  right: 0px;
 }
 
 .slider-prev-btn {
-  left: 8px;
+  left: 0px;
+}
+
+.skeleton {
+  background-color: #dddbdd;
+  color: #dddbdd;
+  background-size: 300%;
+  position: relative;
+  overflow: hidden;
+}
+
+.skeleton:after {
+  -webkit-animation: shimmer 2s infinite;
+  animation: shimmer 2s infinite;
+  background-image: linear-gradient(
+    90deg,
+    hsla(0, 0%, 100%, 0),
+    hsla(0, 0%, 100%, 0.2) 20%,
+    hsla(0, 0%, 100%, 0.5) 60%,
+    hsla(0, 0%, 100%, 0)
+  );
+  bottom: 0;
+  content: "";
+  left: 0;
+  position: absolute;
+  right: 0;
+  top: 0;
+  -webkit-transform: translateX(-100%);
+  transform: translateX(-100%);
+}
+
+.skeleton__hot-collections--img {
+  height: 100%;
+}
+
+.skeleton__hot-collections--author--img {
+  height: 60px;
+  border-radius: 50%;
+}
+
+@keyframes shimmer {
+  0% {
+    transform: translateX(-100%);
+  }
+  100% {
+    transform: translateX(100%);
+  }
 }
